Move ConfigHelper from oak-run-commons to oak-core Minor refactorings and code cleanups to indexing modules.
