+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/index/ConfigHelper.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.jackrabbit.oak.index.indexer.document.flatfile.pipelined;
+package org.apache.jackrabbit.oak.plugins.index;
 
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
