+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/index/progress/IndexingProgressReporter.java
@@ -91,7 +91,7 @@ public class IndexingProgressReporter implements NodeTraversalCallback {
     }
 
     public void traversedNode(PathSource pathSource) throws CommitFailedException {
-        if (++traversalCount % 10000 == 0) {
+        if (++traversalCount % 100_000 == 0) {
             double rate = traversalRateEstimator.getNodesTraversedPerSecond();
             String formattedRate = String.format("%1.2f nodes/s, %1.2f nodes/hr", rate, rate * 3600);
             String estimate = estimatePendingTraversal(rate);
