+++ oak-lucene/src/main/java/org/apache/jackrabbit/oak/plugins/index/lucene/writer/LuceneIndexWriterConfig.java
@@ -19,9 +19,13 @@
 
 package org.apache.jackrabbit.oak.plugins.index.lucene.writer;
 
+import org.apache.jackrabbit.oak.commons.properties.SystemPropertySupplier;
 import org.apache.lucene.index.IndexWriterConfig;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class LuceneIndexWriterConfig {
+    private final static Logger LOG = LoggerFactory.getLogger(LuceneIndexWriterConfig.class);
     /**
      * This property will be used to set Lucene's IndexWriter.maxBufferedDeleteTerms
      * IndexWriter.maxBufferedDeleteTerms is used to flush buffered data to lucene index.
@@ -34,11 +38,14 @@ public class LuceneIndexWriterConfig {
      */
     public final static String RAM_PER_THREAD_HARD_LIMIT_MB_KEY = "oak.index.lucene.ramPerThreadHardLimitMB";
 
+    private final int maxBufferedDeleteTerms = SystemPropertySupplier.create(
+                    MAX_BUFFERED_DELETE_TERMS_KEY, IndexWriterConfig.DISABLE_AUTO_FLUSH)
+            .loggingTo(LOG).get();
+    private final int ramPerThreadHardLimitMB = SystemPropertySupplier.create(
+                    RAM_PER_THREAD_HARD_LIMIT_MB_KEY, IndexWriterConfig.DEFAULT_RAM_PER_THREAD_HARD_LIMIT_MB)
+            .loggingTo(LOG).get();
+
     private final double ramBufferSizeMB;
-    private final int maxBufferedDeleteTerms = Integer.getInteger(MAX_BUFFERED_DELETE_TERMS_KEY,
-            IndexWriterConfig.DISABLE_AUTO_FLUSH);
-    private final int ramPerThreadHardLimitMB = Integer.getInteger(RAM_PER_THREAD_HARD_LIMIT_MB_KEY,
-            IndexWriterConfig.DEFAULT_RAM_PER_THREAD_HARD_LIMIT_MB);
     private final int threadCount;
 
     public LuceneIndexWriterConfig() {
